refactor to move components into their own folder
