+++ components/bolt.js
@@ -1,7 +1,7 @@
 import React from 'react';
 //import { Image } from 'cross-country';
 import Image from 'next/image';
-import bolt from './bolt.svg';
+// import bolt from './bolt.svg';
 
 const Bolt = () =>  <Image src="/bolt.svg" alt="lightening bolt" width={30} height={30} />;
